# select: Check if CSV is valid utf8