## Description

Check if first block of CSV data is valid utf8.

Fixes #10970

## How to test this PR?

Test with binary files, etc.

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)