@vadmeste 2020-11-27
Something like this seems to be enough:
```
diff --git a/pkg/csvparser/reader.go b/pkg/csvparser/reader.go
index c5d75c128..9bbc554a1 100644
--- a/pkg/csvparser/reader.go
+++ b/pkg/csvparser/reader.go
@@ -254,6 +254,9 @@ func (r *Reader) readLine() ([]byte, error) {
                line[n-2] = '\n'
                line = line[:n-1]
        }
+       if !utf8.Valid(line) {
+               return nil, errors.New("invalid utf8 character")
+       }
        return line, err
 }
```

It doesn't seem to make any difference in benchmark numbers:
```
name                old time/op    new time/op    delta
ReadLargeFields-24    6.59µs ± 0%    6.88µs ± 0%   ~     (p=1.000 n=1+1)

name                old alloc/op   new alloc/op   delta
ReadLargeFields-24    3.94kB ± 0%    3.94kB ± 0%   ~     (all equal)

name                old allocs/op  new allocs/op  delta
ReadLargeFields-24      24.0 ± 0%      24.0 ± 0%   ~     (all equal)
```