@klauspost 2020-11-30
@vadmeste @harshavardhana It is a >4 % slowdown which will be paid by the entire file. That is why I only do it for the first block.

You need more than n=1 for it to show the difference. Use `benchcmp`.