@vadmeste 2020-12-01
> It is a >4 % slowdown which will be paid by the entire file.

yes, I thought it is not a big deal if we want to be complying to S3 behavior here, S3 detects if there is non utf8 sequence even if the file is large. Approving anyway for now.